Remove legacy solver classes.
